+++ Mods ChoGGi/Solar Array Follows Sun/metadata.lua
@@ -1,0 +1,16 @@
+return PlaceObj("ModDef", {
+	"title", "Solar Array Follows Sun v0.1",
+	"version", 1,
+	"saved", 1542801600,
+	"image", "Preview.png",
+	"id", "ChoGGi_SolarArrayFollowsSun",
+	"steam_id", "1570918489",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[The panels do it, why not the array panels?
+
+Requested by tresslessone.]],
+})